requirements.txt
PyYAML>=6.0
markdown>=3.4
jinja2>=3.1.0
click>=8.0.0
rich>=13.0.0
python-dateutil>=2.8.0
pathlib>=1.0.0
typing-extensions>=4.0.0